Incomplete refactoring of B/C command
Committing this as a snapshot of a refactoring in process... splitting the B/C command into smaller pieces.
Most significant change here is that `imageDisplay.changeTimestampProperty()` is used within `QuPathViewer` to detect when repaint requests are required, and these are queued rather than processed immediately.

This greatly reduces the requirement to explicitly call `QuPathViewer.repaintEntireImage()`, because the viewer is notified via invalidation of the timestamp that a repaint is required.

I'm hoping this leads to simpler, more maintainable and (possibly) more performant code.
