+++ qupath-gui-fx/src/main/java/qupath/lib/gui/actions/CommonActions.java
@@ -30,7 +30,7 @@ import qupath.lib.gui.QuPathGUI;
 import qupath.lib.gui.actions.annotations.ActionAccelerator;
 import qupath.lib.gui.actions.annotations.ActionConfig;
 import qupath.lib.gui.actions.annotations.ActionIcon;
-import qupath.lib.gui.commands.BrightnessContrastCommand;
+import qupath.lib.gui.commands.display.BrightnessContrastCommand;
 import qupath.lib.gui.commands.Commands;
 import qupath.lib.gui.commands.ContextHelpViewer;
 import qupath.lib.gui.commands.CountingPanelCommand;
